H2: Complete Problem 6
Wrote and tested Problem 6 solution.

Problem 6:

A problem that has a user enter positive numbers and keeps record of how many even  numbers are entered. When the user enters 0, the problem stops and then displays the number of even  numbers entered. For example, if a user enters 3, 56, 4, 13, 779, 46 and 0, then the problem would display something like “User entered 3 even numbers”. You need to use a loop structure for this one.