H3: Problem 4 Solution
For income 0 - 15000, apply a 0% State Tax.

For income 15001 - 30000, apply a 5% State Tax.

For income over 30000, apply a 10% state tax.

Tax brackets stack together.

Problem text below:

Going to find the amount of state income tax that will be paid based on person’s salary. The rate schedules is as follows:
No tax is paid on the first $15,000 of net income.
A tax of 5% is assessed on each dollar of net income from $15,001 to $30,000.
A tax of 10% is assessed on each dollar of net income over $30,000.
Need to calculate the amount of state tax to be paid based on person’s net income. Output should display as shown below, must use printf to get correct format and full credit. Output turned in needs to test following incomes, $40,000, $12,500 and $28,275.

Please enter the person's net income --> 24500
Net Income = $24500.00
State Tax = $475.00

I will be adding the printf statements in another commit.