H4: Problem 1 Solution
Problem 1:

1) For this problem you will write a program and an algorithm that determines a person’s GPA,
this will be based on taking only two classes. The GPA is found by using following formula:

GPA = Total Points / Total Number of Credits

The Total Points is found from taking the number of credits times the number value for
corresponding letter grade and adding both classes together. The letter
grades have following:
A = 4.0 B = 3.0 C = 2.0 D = 1.0 F = 0.0

We will not use and plus or minus grades, and we will assume user enters a capital letter only.
The output will display the letter grade and number of credits for both classes. Once the GPA is
determined its value will be displayed, and if the GPA is less than 2.0 a message will say “You
are doing poorly” and if the GPA is 3.5 or above the message will display “Congratulations,
doing good”. Below is sample output, test program with this and also B,3 A,4 and then with
D,3 C,3.

Enter letter grade for first class --> B
Enter number of credit hours for first class --> 3

Enter letter grade for second class --> C
Enter number of credit hours for second class --> 4

B 3
C 4
Your GPA = 2.42857