H4: Problem 2 Solution
Problem 2 text:

For this problem turn in both the algorithm and source code. Need to determine if you are
legally intoxicated or not, based on chart given below, use the formula for blood alcohol content:

BAC = ((Alcohol consumed in grams / (Body weight in grams * R)) *100) – Metabolism Rate

 R = Gender, a constant in this equation (.55 for females and .68 for males)
 Grams of Alcohol Consumed = Number of drinks * 14
 Body weight in grams = Body weight in pounds * 454
 Metabolism Rate = Number hours drinking * 0.015

Example
180lb male who has drank 4 beers over the course of 2 hours will have the following BAC level.

((4 beers *14)/ ((180×454) * .68)) * 100) - (2 * 0.015) = 0.07 BAC

Once the BAC is calculated, display the BAC and then use this table below to display
appropriate message. Need to display with at least 3 to 4 decimal places as well.