H5: Add Lab 5 solution
Lab 5 text:

Write a program that will ask the user to enter any integer, the program will then tell the
user if the number is Positive, Negative or Zero. It will also keep count of how many of
each you have entered. Need to keep asking and doing this until user enters an ‘N’. When
user chooses to stop, it will display the number of Positive, Negative and Zeros entered.