H6: Problem 2 solution
Provide problem 2 solution.

Problem text is as follows:

Write a program that has the user enter two integers and then has the user enter a math operand to do the mathematic operation on the two inputted integers. The operands are ‘+’, ‘-’, ‘*’, ‘/’, and ‘%’, any other character entered for the math operand should produce an error message of some kind. If the operand is correct, then the mathematical output should be displayed along with the correct formula. For example, if 4 and 6 and + are entered, output should be something like the following: 4 + 6 = 10. To receive full credit you must use a “switch” statement in your program. Program should also continue to loop until the user chooses ‘N’ for No in some kind of loop.