H7: Problem 2 solution
Problem 2 text:

Program will have user enter an integer between 1 and 100, anything else
will cause the loop and program to quit. This integer is first tested to see if it is an odd or
even number; here an appropriate message is displayed, displayed only in the main
function. Then the sum of all the numbers up to and including the entered number is
calculated and displayed. So if the user enters the number 4, the sum of numbers from 1
to 4 would yield 10. This program needs to call two separate methods, one to determine
if number is odd or even, and the other to calculate the sum of the numbers. Both
methods will have one parameter and both should return something back, which is the
answer the main will deal with. Remember to design your functions so that they are
logically coherent and follow the guidelines given in class and examples.