H7: Lab 7 solution
Write a program that continually ask the user to enter any two integers and then subtracts the
second from the first. The answer is displayed along with a symbol for positive or negative. So
if the result is a positive number then a large PLUS sign is displayed, if result is negative then a
large NEGATIVE sign is displayed. Each sign will be displayed by using a method/function to
draw out the symbol. Both functions will have no parameters and return nothing back.
Program will continue until user enters both numbers that are the same, should probably use a
do-while loop, and display a PLUS sign if the result is zero.