H8: Lab9 Initial Solution
Lab Goals – 1) Learning how to write functions.
2) Writing functions that return information back via the function name.

Write a program that has user enter three integers, the largest of the three will then be displayed.

Use a function that will accept the three integers from the main function and then calculate the maximum of those three and pass it back.

You also must have a loop structure of some kind that allows the user to continue until they press ‘N’ or ‘n’ to stop.

Use an if –else if structure to determine which number is the largest.